initial plot_forest on bokeh
